feat: Add Social icons for Google and Facebook login
This commit adds two new icons, GoogleIcon and FacebookIcon, to the codebase. These icons will be used for the Google and Facebook login buttons in the login screen. The icons are implemented using the @gluestack-ui/icon package and styled with the @gluestack-ui/themed package. This addition enhances the user experience by providing visually appealing icons for social login options.