feat: Update User struct to include ExpoPushToken field
The User struct in database.go has been updated to include a new field called ExpoPushToken. This field will store the Expo push token for a user, allowing for push notifications to be sent to specific users. This change was made to support the new functionality of retrieving push tokens for specific users.