refactor: Update user details in LoginForm and OtpPage
This commit updates the LoginForm and OtpPage components to correctly set the user details based on the form input. In LoginForm, the email field is used to set the user email in the setUserDetails function. In OtpPage, the email field is used to fetch the user details from the API and set them in the setUserDetails function. These changes ensure that the correct user details are stored and used throughout the application.