feat: Update user details in AuthService
The `updateUserDetails` function in `AuthService.ts` has been added to handle updating user details. This function makes a POST request to the API endpoint `/update-user` with the provided user details. If the response status is 200, it returns the response data. Otherwise, it throws an error with the response message or a default error message. If an unexpected error occurs, it throws an error with a generic message.