feat: Add user creation and IP management endpoints
This commit adds new endpoints for creating users and managing IP addresses. The `CreateUser` function is responsible for creating a new user in the database, while the `AddIP` and `RemoveIP` functions handle adding and removing IP addresses for a user respectively. These changes enhance the functionality of the application by providing the ability to manage user accounts and their associated IP addresses.